+++ openeo_driver/datacube.py
@@ -100,7 +100,13 @@ class DriverDataCube:
         self._not_implemented()
 
     def apply_neighborhood(
-        self, process: dict, *, size: List[dict], overlap: List[dict], context: Optional[dict] = None, env: EvalEnv
+        self,
+        process: dict,
+        *,
+        size: List[dict],
+        overlap: Optional[List[dict]] = None,
+        context: Optional[dict] = None,
+        env: EvalEnv,
     ) -> "DriverDataCube":
         self._not_implemented()