+++ CHANGELOG.md
@@ -19,7 +19,10 @@ and start a new "In Progress" section above it.
 
 <!-- start-of-changelog -->
 
-## In progress: 0.134.0
+## In progress: 0.135.0
+
+
+## 0.134.0
 
 - Introduce `asset_url` option to allow backend implementations to have custom code for retrieving assets. Default
   behavior remains unchanged.